fix: split ui-coverage-runner tests to satisfy line-count lint
